Consolidate protocols *always*
The reason protcol consolidation was disabled for test env was because
some protocol implementations were defined in tests.

These are now defined under lib/ so and now we can use consolidation.

Tests now run in 0.1-0.2 seconds on my MBP.
