+++ mix.exs
@@ -17,7 +17,7 @@ defmodule GraphQL.Mixfile do
      homepage_url: @repo_url,
      build_embedded: Mix.env == :prod,
      start_permanent: Mix.env == :prod,
-     consolidate_protocols: Mix.env == :prod,
+     consolidate_protocols: true,
      name: "GraphQL",
      docs: [main: "GraphQL", logo: "logo.png", extras: ["README.md"]]]
   end
